v0.0.4

+++ spec/graphoid/queries/relations/has_one_through_spec.rb
@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 require 'rails_helper'
 
-describe "QueryHasOneThrough", :type => :request do
-  next #if ENV['DRIVER'] == "mongo"
+describe 'QueryHasOneThrough', type: :request do
+  next # if ENV['DRIVER'] == "mongo"
   # TODO: make "through" relations not always be referenced to many.
 
   let!(:delete) { Brain.delete_all; Person.delete_all; Account.delete_all }
-  subject { Helper.resolve(self, "accounts", @query) }
+  subject { Helper.resolve(self, 'accounts', @query) }
 
   let!(:a0) { Account.create! }
   let!(:a1) { Account.create! }
@@ -19,9 +21,8 @@ describe "QueryHasOneThrough", :type => :request do
   let!(:b1) { Brain.create!(name: 'b1', person: p1) }
   let!(:b2) { Brain.create!(name: 'c0', person: p2) }
 
-  describe "filtering with conditions in has_one_through relations" do
-
-    it "filters by has_one relation" do
+  describe 'filtering with conditions in has_one_through relations' do
+    it 'filters by has_one relation' do
       @query = %{
         query {
           accounts(where: {
@@ -37,11 +38,11 @@ describe "QueryHasOneThrough", :type => :request do
 
       expect(subject.size).to eq(2)
 
-      expect(subject[0]["id"]).to eq a0.id.to_s
-      expect(subject[0]["brain"]["id"]).to eq b0.id.to_s
+      expect(subject[0]['id']).to eq a0.id.to_s
+      expect(subject[0]['brain']['id']).to eq b0.id.to_s
 
-      expect(subject[1]["id"]).to eq a1.id.to_s
-      expect(subject[1]["brain"]["id"]).to eq b1.id.to_s
+      expect(subject[1]['id']).to eq a1.id.to_s
+      expect(subject[1]['brain']['id']).to eq b1.id.to_s
     end
   end
 end
